backout 5876 sys/regset.h pollutes name space backout commit 21227944c2bcc086121a5428f3f9d2496ba646f5.
