+++ usr/src/cmd/mdb/sparc/kmdb/kmdb_makecontext.c
@@ -24,6 +24,8 @@
  * Use is subject to license terms.
  */
 
+#pragma ident	"%Z%%M%	%I%	%E% SMI"
+
 /*
  * Context-saving routine used for pipelines.  Designed for use only
  * with kmdb_setcontext, and with the assumption that func() will never
@@ -34,7 +36,6 @@
 #include <ucontext.h>
 #include <sys/types.h>
 #include <sys/stack.h>
-#include <sys/regset.h>
 
 #include <kmdb/kmdb_context_impl.h>
 #include <mdb/mdb_kreg.h>
