+++ usr/src/lib/libc/amd64/gen/makectxt.c
@@ -27,13 +27,14 @@
 /*	Copyright (c) 1988 AT&T	*/
 /*	  All Rights Reserved  	*/
 
+#pragma ident	"%Z%%M%	%I%	%E% SMI"
+
 #pragma weak _makecontext = makecontext
 
 #include "lint.h"
 #include <stdarg.h>
 #include <ucontext.h>
 #include <sys/stack.h>
-#include <sys/regset.h>
 
 /*
  * The ucontext_t that the user passes in must have been primed with a
