+++ usr/src/lib/libc/i386/gen/siglongjmp.c
@@ -28,12 +28,13 @@
 /*	Copyright (c) 1988 AT&T	*/
 /*	  All Rights Reserved  	*/
 
+#pragma ident	"%Z%%M%	%I%	%E% SMI"
+
 #pragma weak _siglongjmp = siglongjmp
 
 #include "lint.h"
 #include <sys/types.h>
 #include <sys/ucontext.h>
-#include <sys/regset.h>
 #include <setjmp.h>
 #include <ucontext.h>
