+++ usr/src/lib/libm/common/m9x/__fex_i386.c
@@ -40,7 +40,6 @@
 #include <sunmath.h>
 #endif
 #include <fenv.h>
-#include <sys/regset.h>
 #include "fex_handler.h"
 #include "fenv_inlines.h"
