+++ usr/src/lib/libm/common/m9x/__fex_sse.c
@@ -34,7 +34,6 @@
 #else
 #include <sys/ieeefp.h>
 #endif
-#include <sys/regset.h>
 #include "fex_handler.h"
 #include "fenv_inlines.h"
