+++ usr/src/lib/libm/common/m9x/fex_log.c
@@ -39,7 +39,6 @@
 #include <string.h>
 #include <signal.h>
 #include <ucontext.h>
-#include <sys/regset.h>
 #include <sys/frame.h>
 #include <fenv.h>
 #include <sys/ieeefp.h>
