+++ usr/src/lib/libproc/common/P32ton.c
@@ -24,6 +24,8 @@
  * Use is subject to license terms.
  */
 
+#pragma ident	"%Z%%M%	%I%	%E% SMI"
+
 #include <sys/types.h>
 #include <sys/mkdev.h>
 #include <sys/regset.h>
@@ -265,7 +267,7 @@ prfpregset_32_to_n(const prfpregset32_t *src, prfpregset_t *dst)
 #elif defined(__amd64)
 
 	struct _fpstate32 *src32 = (struct _fpstate32 *)src;
-	struct _fpchip_state *dst64 = (struct _fpchip_state *)dst;
+	struct fpchip_state *dst64 = (struct fpchip_state *)dst;
 	int i;
 
 	(void) memcpy(dst64->st, src32->_st, sizeof (src32->_st));
@@ -596,7 +598,7 @@ prfpregset_n_to_32(const prfpregset_t *src, prfpregset32_t *dst)
 #elif defined(__amd64)
 
 	struct _fpstate32 *dst32 = (struct _fpstate32 *)dst;
-	struct _fpchip_state *src64 = (struct _fpchip_state *)src;
+	struct fpchip_state *src64 = (struct fpchip_state *)src;
 	uint32_t top;
 	int i;
