+++ usr/src/uts/common/sys/core.h
@@ -20,14 +20,14 @@
  * CDDL HEADER END
  */
 /*
- * Copyright 2015 Nexenta Systems, Inc.  All rights reserved.
- *
  * Copyright 2004 Sun Microsystems, Inc.  All rights reserved.
  * Use is subject to license terms.
  */
 
 #ifndef	_SYS_CORE_H
 #define	_SYS_CORE_H
+
+#pragma ident	"%Z%%M%	%I%	%E% SMI"
 
 #ifndef _KERNEL
 #include <sys/reg.h>
@@ -63,9 +63,9 @@ struct core {
 	int	c_dsize;		/* Data size (bytes) */
 	int	c_ssize;		/* Stack size (bytes) */
 	char	c_cmdname[CORE_NAMELEN + 1]; /* Command name */
-	struct	_fpu	c_fpu;		/* external FPU state */
+	struct	fpu c_fpu;		/* external FPU state */
 #if defined(__sparc)
-	struct	_fq c_fpu_q[_MAXFPQ];	/* fpu exception queue */
+	struct	fq c_fpu_q[MAXFPQ];	/* fpu exception queue */
 #endif
 	int	c_ucode;		/* Exception no. from u_code */
 };
