+++ usr/src/uts/sparc/sys/procfs_isa.h
@@ -27,6 +27,8 @@
 #ifndef _SYS_PROCFS_ISA_H
 #define	_SYS_PROCFS_ISA_H
 
+#pragma ident	"%Z%%M%	%I%	%E% SMI"
+
 /*
  * Instruction Set Architecture specific component of <sys/procfs.h>
  * sparc v8/v9 version
@@ -167,7 +169,7 @@ typedef struct prfpregset {
 	uint8_t	pr_q_entrysize;		/* # of bytes per FQ entry */
 	uint8_t	pr_en;			/* flag signifying fpu in use */
 	char	pr_pad[13];		/* ensure sizeof(prfpregset)%16 == 0 */
-	struct _fq pr_q[16];		/* contains the FQ array */
+	struct fq pr_q[16];		/* contains the FQ array */
 } prfpregset_t;
 #else
 typedef struct prfpregset {
@@ -180,7 +182,7 @@ typedef struct prfpregset {
 	uint8_t	pr_qcnt;		/* # of entries in saved FQ */
 	uint8_t	pr_q_entrysize;		/* # of bytes per FQ entry */
 	uint8_t	pr_en;			/* flag signifying fpu in use */
-	struct _fq pr_q[32];		/* contains the FQ array */
+	struct fq pr_q[32];		/* contains the FQ array */
 } prfpregset_t;
 #endif	/* __sparcv9 */
